+++ src/onboarding/WelcomeContainer.js
@@ -2,8 +2,8 @@ import { connect } from 'react-redux';
 import Hello from './Hello';
 import { pushProfilePictureRoute } from 'src/navigation/actions';
 
-import { getProfileId } from 'src/auth/reducer';
-import { getProfile } from 'src/profile/reducer';
+import { getProfileId } from 'src/reducers/auth';
+import { getProfile } from 'src/reducers/profile';
 
 export default connect(
   state => ({
