+++ src/reducers/index.js
@@ -1,9 +1,9 @@
 import { combineReducers } from 'redux';
 import entities from 'src/reducers/entities';
-import auth from 'src/auth/reducer';
-import navigation from 'src/navigation/reducer';
-import profile from 'src/profile/reducer';
-import onboarding from 'src/onboarding/reducer';
+import auth from 'src/reducers/auth';
+import navigation from 'src/reducers/navigation';
+import profile from 'src/reducers/profile';
+import onboarding from 'src/reducers/onboarding';
 
 
 const rootReducer = combineReducers({
